Refactor Thread Register Access and Improve Windows Backend
- Simplified register access methods in Thread class by utilizing a unified interface in the Registers class.
- Updated Backend class to accept a pointer to Arch instead of a value, enhancing flexibility.
- Changed attach method signature in WindowsBackend to use uintptr_t for process IDs.
- Enhanced error handling and logging in WindowsBackend during process attachment.
- Introduced helper methods for handling suspended threads in WindowsBackend.
- Improved register capture methods for both native and Wow64 processes in WindowsBackend.
- Added support for enumerating symbols in DbgHelpBackend, allowing for more flexible symbol lookups.
- Updated symbol loading logic to allow mismatched PDBs for local development.

+++ include/smalldbg/Process.h
@@ -14,11 +14,11 @@ class Debugger;
 
 class Process {
 public:
-    Process(Debugger* dbg, int pid);
+    Process(Debugger* dbg, uintptr_t pid);
     ~Process() = default;
 
     // Process identification
-    int getPid() const { return pid; }
+    uintptr_t getPid() const { return pid; }
     
     // Thread enumeration
     std::vector<std::shared_ptr<Thread>> threads();
@@ -42,7 +42,7 @@ public:
 
 private:
     Debugger* debugger{nullptr};
-    int pid;
+    uintptr_t pid;
     std::map<ThreadId, std::shared_ptr<Thread>> threadMap;
     std::shared_ptr<Thread> firstThread;
 };
